+++ examples/complex.py
@@ -1,14 +1,15 @@
 import asyncio
 
 from tiny_listener import Listener, Context
+from concurrent.futures import TimeoutError
 
 
 class App(Listener):
     async def listen(self, todo):
         todo("step_1", cid="Alice")
         todo("step_3", cid="Bob")
-        todo("step_2", cid="Bob")
-        todo("step_3", cid="Alice", parents_timeout=1)
+        todo("step_2", cid="Bob", timeout=1)
+        todo("step_3", cid="Alice")
         todo("step_1", cid="Bob")
         todo("step_2", cid="Alice")
 
@@ -16,9 +17,9 @@ class App(Listener):
 app = App()
 
 
-@app.error_raise
-async def err(ctx: Context):
-    print(ctx.errors)
+@app.error_raise(TimeoutError)
+async def err_handler(err: TimeoutError):
+    print(err)
 
 
 @app.do("step_1")