+++ tests/test_context.py
@@ -1,7 +1,7 @@
 import pytest
 from unittest import TestCase
 
-from tiny_listener import Context, Listener, Event
+from tiny_listener import Context, Listener, Event, Route, NotFound
 
 
 class TestContext(TestCase):
@@ -63,6 +63,14 @@ class TestContext(TestCase):
         event = ctx.new_event("/event/foo")
         self.assertEqual(event, ctx.events["/event/foo"])
 
+    def test_todo(self):
+        class App(Listener):
+            def listen(self, _): ...
+
+        app = App()
+        ctx = app.new_context("ctx_11")
+        self.assertRaises(NotFound, ctx.todo, "something")
+
 
 class TestEvent(TestCase):
     def setUp(self) -> None:
@@ -83,15 +91,9 @@ class TestEvent(TestCase):
     def test_ctx(self):
         self.assertIs(self.ctx, self.event_foo.ctx)
 
-    def test_load_parents(self):
-        event = Event(name="foo", ctx=self.ctx)
-        self.assertFalse(event.parents_ready.is_set())
-        event.parents_count = 0
-        self.assertFalse(event.parents_ready.is_set())
-        event.load_parents()
-        self.assertTrue(event.parents_ready.is_set())
+    def test_parents(self):
+        async def _cb(): ...
 
-    def test_add_parents(self):
         # match all
         for pats in [
             [""],
@@ -101,14 +103,14 @@ class TestEvent(TestCase):
             ["/user/foo", "/user/bar"],
             ["", "/user/_not_exist_"],
         ]:
-            event = Event(name="foo", ctx=self.ctx).add_parents(*pats)
+            event = Event(name="foo", ctx=self.ctx, route=Route("/", _cb, parents=pats))
             self.assertEqual({self.event_foo, self.event_bar}, event.parents)
         # match one
         for pats in [
             ["/user/foo"],
             ["/user/foo", "/user/f"]
         ]:
-            event = Event(name="foo", ctx=self.ctx).add_parents(*pats)
+            event = Event(name="foo", ctx=self.ctx, route=Route("/", _cb, parents=pats))
             self.assertEqual({self.event_foo}, event.parents)
         # match none
         for pats in [
@@ -116,15 +118,8 @@ class TestEvent(TestCase):
             ["/user/_not_exist_"],
             ["/user/_not_exist_foo", "/user/_not_exist_bar"],
         ]:
-            event = Event(name="foo", ctx=self.ctx).add_parents(*pats)
+            event = Event(name="foo", ctx=self.ctx, route=Route("/", _cb, parents=pats))
             self.assertEqual(set(), event.parents)
-
-    def test_set_data(self):
-        self.event_foo.set_data({"field_A": 1})
-        self.assertEqual({"field_A": 1}, self.event_foo.data)
-
-        self.event_foo.set_data({"field_A": 2, "field_B": 3})
-        self.assertEqual({"field_A": 2, "field_B": 3}, self.event_foo.data)
 
 
 @pytest.mark.asyncio
@@ -132,12 +127,12 @@ async def test_event_trigger(event_loop):
     class App(Listener):
         def listen(self, _): ...
 
+    async def _cb(): ...
+
     app = App()
     ctx = app.new_context("ctx_trigger")
     event_foo = ctx.new_event("/user/foo")
-    event_bar = ctx.new_event("/user/bar")
-    event_bar.parents_count = 1
-    event_bar.add_parents("/user/foo")
+    event_bar = ctx.new_event("/user/bar", route=Route("/", fn=_cb, parents=["/user/foo"]))
 
     assert event_foo.is_done is False
     event_loop.call_later(0.1, lambda: event_foo.done())