+++ tiny_listener/listener.py
@@ -1,6 +1,6 @@
 import asyncio
 import signal
-from typing import Optional, Dict, Callable, List, Any, Union, Coroutine
+from typing import Optional, Dict, Callable, List, Any, Union, Coroutine, Tuple, Type
 
 from .context import Context
 from .routing import Route, _EventHandler, EventHandler, as_handler, Params
@@ -21,7 +21,7 @@ class Listener:
 
         self._pre_do: List[EventHandler] = []
         self._post_do: List[EventHandler] = []
-        self._error_raise: List[EventHandler] = []
+        self._error_raise: List[Tuple[EventHandler, Type[BaseException]]] = []
 
     def new_context(self, cid: Optional[str] = None, **scope: Any) -> Context:
         return Context(cid, _listener_=self, **scope)
@@ -30,21 +30,23 @@ class Listener:
         for t in asyncio.Task.all_tasks(self.loop):
             t.cancel()
 
-    def pre_do(self, handler: _EventHandler) -> None:
-        self._pre_do.append(as_handler(handler))
+    def pre_do(self, fn: _EventHandler) -> None:
+        self._pre_do.append(as_handler(fn))
 
-    def post_do(self, handler: _EventHandler) -> None:
-        self._post_do.append(as_handler(handler))
+    def post_do(self, fn: _EventHandler) -> None:
+        self._post_do.append(as_handler(fn))
 
-    def error_raise(self, handler: _EventHandler) -> None:
-        self._error_raise.append(as_handler(handler))
+    def error_raise(self, exc: Type[BaseException]) -> Callable:
+        def f(fn: _EventHandler) -> Any:
+            self._error_raise.append((as_handler(fn), exc))
+        return f
 
     def todo(
             self,
             name: str,
             cid: Optional[str] = None,
             block: bool = False,
-            parents_timeout: Optional[float] = None,
+            timeout: Optional[float] = None,
             data: Optional[Dict] = None
     ) -> Coroutine or None:
         route = None
@@ -58,27 +60,22 @@ class Listener:
             raise NotFound(f"route `{name}` not found")
 
         params = Params(params)
-
         ctx = self.new_context(cid)
-        event = ctx.new_event(name)
-        event.parents_count = route.opts.get("parents_count")
-        event.add_parents(*route.opts["parents"]).set_data(data or {})
-        event.parents_timeout = parents_timeout
+        event = ctx.new_event(name=name, timeout=timeout, route=route, **data or {})
 
         async def _todo():
             async with event as exc:
                 try:
                     if exc:
                         raise exc
-                    [await fn(ctx, event, params) for fn in self._pre_do]
+                    [await fn(ctx, event, params, exc) for fn in self._pre_do]
                     res = await route.fn(ctx, event, params)
-                    [await fn(ctx, event, params) for fn in self._post_do]
+                    [await fn(ctx, event, params, exc) for fn in self._post_do]
                     return res
                 except BaseException as e:
                     if not self._error_raise:
                         raise e
-                    ctx.errors.append(e)
-                    [await fn(ctx, event, params) for fn in self._error_raise]
+                    [await fn(ctx, event, params, e) for fn, exc_cls in self._error_raise if isinstance(e, exc_cls)]
 
         if block:
             return _todo()
@@ -96,27 +93,25 @@ class Listener:
             self,
             path: str,
             parents: Union[None, List[str], Callable[[Context], List[str]]] = None,
-            parents_count: Optional[int] = None,
             **kwargs: Any
     ) -> Callable[[_EventHandler], None]:
         parents = parents or []
 
         def _decorator(fn: _EventHandler) -> None:
-            self.routes.append(Route(path=path, fn=fn, opts={"parents": parents, "parents_count": parents_count, **kwargs}))
+            self.routes.append(Route(path=path, fn=fn, parents=parents, opts=kwargs))
         return _decorator
 
     @classmethod
-    def default_route(
+    def default_do(
             cls,
             path: str,
             parents: Union[None, List[str], Callable[[Context], List[str]]] = None,
-            parents_count: Optional[int] = None,
             **kwargs: Any
     ) -> Callable[[_EventHandler], None]:
         parents = parents or []
 
         def _decorator(fn: _EventHandler) -> None:
-            cls.__default_routes__.append(Route(path=path, fn=fn, opts={"parents": parents, "parents_count": parents_count, **kwargs}))
+            cls.__default_routes__.append(Route(path=path, fn=fn, parents=parents, opts=kwargs))
         return _decorator
 
     def __repr__(self) -> str: