+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Change Log
 
+## 0.0.6 (2021-11-03)
+
+### Changed
+
+- Remove parent count
+- Refactor route parents
+- Refactor error raise handler
+
+
 ## 0.0.5 (2021-10-20)
 
 ### Changed